feat: implement useDashboardGuard hook for token validation and redirect logic
