+++ turnomaster-project/resources/js/routers/DashboardRouter.tsx
@@ -1,38 +1,12 @@
-import React, { JSX } from 'react';
-import { Routes, Route, Navigate, useLocation } from 'react-router-dom';
+import React from 'react';
+import { Routes, Route, useLocation } from 'react-router-dom';
 import { AnimatePresence, motion } from 'framer-motion';
 import DashboardLayout from '../Layouts/Dashboard/DashboardLayout';
+import { useDashboardGuard } from '../hooks/useDashboardGuard';
 
 import Index from '../Pages/Dashboard/Index/Index';
 import Profile from '../Pages/Dashboard/Profile/Profile';
 import Settings from '../Pages/Dashboard/Settings/Settings';
-
-function parseJwt(token: string): { exp?: number } | null {
-    try {
-        const base64Payload = token.split('.')[1];
-        const payload = atob(base64Payload);
-        return JSON.parse(payload);
-    } catch (error) {
-        console.error('Failed to parse JWT:', error);
-        return null;
-    }
-}
-
-function DashboardGuard({ children }: { children: JSX.Element }) {
-    const token = localStorage.getItem('token');
-
-    if (!token) {
-        return <Navigate to="/auth/login" />;
-    }
-
-    const decoded = parseJwt(token);
-    if (!decoded || !decoded.exp || decoded.exp * 1000 < Date.now()) {
-        localStorage.removeItem('token');
-        return <Navigate to="/auth/login" />;
-    }
-
-    return children;
-}
 
 function DashboardRouter() {
     const location = useLocation();
@@ -46,7 +20,7 @@ function DashboardRouter() {
 
     return (
         <DashboardLayout>
-            <DashboardGuard>
+            {useDashboardGuard(
                 <AnimatePresence mode="wait">
                     <motion.div key={location.pathname} {...pageTransition} className="w-full">
                         <Routes location={location} key={location.pathname}>
@@ -56,7 +30,7 @@ function DashboardRouter() {
                         </Routes>
                     </motion.div>
                 </AnimatePresence>
-            </DashboardGuard>
+            )}
         </DashboardLayout>
     );
 }
